third_party/WebKit/LayoutTests/inspector-enabled/sources/debugger/script-formatter-console-expected.txt

CONSOLE ERROR: line 14: Test message.
Tests that the script formatting changes console line numbers.


Running: testSetup

Running: testConsoleMessagesForFormattedScripts
script-formatter-console.html:14 
Pre-format row message list:
[13]
script-formatter-con…e.html:formatted:14 
Post-format row message list:
[13]
